+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Transform OData queries to SQLAlchemy expressions with the new
+  AstToSqlAlchemyClauseVisitor.
+
 ## [0.1.0] - 2021-03-12
 
 ### Added